creating a layout and some validations

+++ src/main/java/src/view/validations/payment/PricePerQuantityExpression.java
@@ -2,7 +2,7 @@ package src.view.validations.payment;
 
 public class PricePerQuantityExpression {
 
-    public static Double calculusResult (Double price, Double quantity) {
+    public static Double calculusResult (Double price, Integer quantity) {
 
         return price * quantity;
